+++ src/panoptes/__init__.py
@@ -1,2 +1,1 @@
-# -*- coding: utf-8 -*-
 __import__('pkg_resources').declare_namespace(__name__)
