+++ tests/test_serializers.py
@@ -1,10 +1,8 @@
 import pytest
-
 from astropy import units as u
-
-from panoptes.utils import current_time
-from panoptes.utils import serializers
 from panoptes.utils import error
+from panoptes.utils import serializers
+from panoptes.utils.time import current_time
 
 
 @pytest.fixture(scope='function')
